mini two functions need to two file
